+++ lib/src/state_machine/mp_th2_file_edit_state_machine/mp_th2_file_edit_state_select_empty_selection.dart
@@ -69,6 +69,13 @@ class MPTH2FileEditStateSelectEmptySelection extends MPTH2FileEditState
           // openOptionEditOverlayWindow();
           keyProcessed = true;
         }
+      case LogicalKeyboardKey.keyV:
+        if ((isCtrlPressed || isMetaPressed) &&
+            !isAltPressed &&
+            !isShiftPressed) {
+          th2FileEditController.elementEditController.pasteElements();
+          keyProcessed = true;
+        }
     }
 
     if (keyProcessed) {